uart: a test for the interrupt, modem lines, and two 32-bit ceilings
**test_uart.c, because a mechanism was added with nothing exercising
it.** The interrupt logic is fiddly for one reason: transmission here
is instantaneous, so THRE is permanently true and an interrupt
conditioned on it alone re-asserts the instant it is acknowledged. The
test pins both halves of the acknowledgement -- reading IIR clears the
transmit indication, writing THR sets it again -- because getting
either wrong gives a UART that works for exactly one byte. It also
pins that receive is *not* cleared by reading IIR (only by taking the
byte), that receive outranks transmit, and that a UART with no line
configured behaves as it did before there was one, which is every
board in this tree. 586 checks, up from 550.

**MSR reported zero, which is an unplugged cable.** A 16550 with no
modem lines tells a driver there is no clear-to-send and no carrier: a
tty opened without CLOCAL blocks on carrier and one with hardware flow
control never starts a transmission. What this device actually is -- a
transmitter that completes instantly into a host that is always ready
-- is a peer holding CTS, DSR and DCD, so that is what it reports now.
It did not fix the userspace-output problem; it is right regardless.

**Two 32-bit ceilings, in the same value, found because a 6e9 cap and
a 5e9 cap both reported an unchanged number.**

emu_run_env_t::max_insn was uint32_t while emu_args_t::max_insn was
uint64_t, and emu_main assigned one to the other through an explicit
cast -- so nothing warned and `--max-insn 6000000000` silently became
1,705,032,704. Every Linux run this session was capped at 1.7e9 while
asking for 6e9.

Then the *report* truncated too: `(unsigned)retired`. With the cap
fixed, a 5e9 run still printed 705,032,704, which reads as the cap
still being wrong rather than the print. Fixed with emu_u64_str, a
formatter rather than `%llu`, because emu_stats.c is shared with the
boards and a newlib built without long-long support prints the format
string instead of the number.

A spinning guest at `--max-insn 5000000000` now reports exactly
5000000000.

+++ src/platform/common/emu_run.h
@@ -34,7 +34,20 @@ typedef enum {
 
 typedef struct emu_run_env {
     uint32_t slice; /* guest instructions between returns to us */
-    uint32_t max_insn; /* 0 = no cap */
+    /*
+     * 0 = no cap.
+     *
+     * **64-bit, because the interesting budgets are above 4G.** This
+     * was uint32_t while emu_args_t::max_insn was uint64_t, and
+     * emu_main assigned one to the other through an explicit cast --
+     * so nothing warned and `--max-insn 6000000000` silently became
+     * 1,705,032,704. It presented as two runs retiring an identical
+     * count, which this project reads as "the code never ran"; here it
+     * meant both had been truncated to the same wrong number. A Linux
+     * kernel reaching userspace is about 1.5e9, so this ceiling sits
+     * exactly where the work is.
+     */
+    uint64_t max_insn;
 
     /*
      * What used to be here: `poll`, `gdb_attached` and `gdb_run`.
